chore: split custom types into types module (#28204)
(cherry picked from commit f7a7cef54a278905c4938823e2d245db5d34efb4)
