+++ frappe/types/frappedict.py
@@ -1,0 +1,20 @@
+class _dict(dict):
+	"""dict like object that exposes keys as attributes"""
+
+	__slots__ = ()
+	__getattr__ = dict.get
+	__setattr__ = dict.__setitem__
+	__delattr__ = dict.__delitem__
+	__setstate__ = dict.update
+
+	def __getstate__(self):
+		return self
+
+	def update(self, *args, **kwargs):
+		"""update and return self -- the missing dict feature in python"""
+
+		super().update(*args, **kwargs)
+		return self
+
+	def copy(self):
+		return _dict(self)
